Disabled dynamic finish_signal_addr allocation for now. It was the root of the bug.

+++ cl_manycore/libraries/bsg_manycore_cuda.h
@@ -42,7 +42,7 @@ typedef struct {
 	char *name;
 	uint32_t argc;
 	uint32_t *argv;
-	uint32_t finish_signal_addr;
+	uintptr_t finish_signal_addr;
 } kernel_t;
 
 typedef struct {
